backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic[email]
python-multipart
python-dotenv
PyJWT
requests
